fix review notes
