+++ src/features/layers_in_area/atoms/areaLayersDetailsResource/index.ts
@@ -1,0 +1,1 @@
+export { areaLayersDetailsResourceAtom } from './areaLayersDetailsResourceAtom';
