+++ src/features/layers_in_area/atoms/areaLayersDetailsResource/types.ts
@@ -1,0 +1,8 @@
+export interface DetailsRequestParams {
+  layersToRetrieveWithGeometryFilter: string[];
+  layersToRetrieveWithoutGeometryFilter: string[];
+  layersToRetrieveWithEventId: string[];
+  geoJSON?: GeoJSON.GeoJSON;
+  eventId?: string;
+  eventFeed?: string;
+}
